requirements.txt
absl-py==0.8.1
appdirs==1.4.4
aspy.yaml==1.3.0
astor==0.8.1
astroid==2.4.1
attrs==19.2.0
audioread==2.1.8
autopep8==1.5.3
Babel==2.8.0
backcall==0.1.0
bce-python-sdk==0.8.53
blackhole==1.0.1
bleach==3.1.0
cachetools==4.0.0
certifi==2019.9.11
cffi==1.14.0
cfgv==2.0.1
chardet==3.0.4
cityscapesScripts==2.2.0
Click==7.0
cloudpickle==1.6.0
cma==2.7.0
colorama==0.4.4
coloredlogs==15.0.1
colorlog==4.1.0
cycler==0.10.0
Cython==0.29
decorator==4.4.2
dill==0.3.3
easydict==1.9
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.2
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Cors==3.0.8
forbiddenfruit==0.1.3
funcsigs==1.0.2
future==0.18.0
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.10.0
google-auth-oauthlib==0.4.1
graphviz==0.13
grpcio==1.35.0
gunicorn==20.0.4
gym==0.12.1
h5py==2.9.0
humanfriendly==9.2
identify==1.4.10
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
importlib-metadata==0.23
ipykernel==5.1.0
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.0
jieba==0.42.1
Jinja2==2.10.3
joblib==0.14.1
JPype1==0.7.2
json5==0.9.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.0
jupyter-lsp==0.8.0
jupyterlab==2.1.3
jupyterlab-server==1.1.5
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
librosa==0.7.2
lightgbm==3.1.1
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
moviepy==1.0.1
multiprocess==0.70.11.1
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.9
networkx==2.4
nltk==3.4.5
nodeenv==1.3.4
notebook==5.7.8
numba==0.48.0
numpy==1.16.4
oauthlib==3.1.0
objgraph==3.4.1
opencv-python==4.1.1.26
openpyxl==3.0.5
packaging==20.9
paddlehub==2.0.4
paddlenlp==2.0.1
paddlepaddle-gpu==2.1.0.post101
paddleseg==2.2.0
pandas==1.1.5
pandocfilters==1.4.2
parl==1.4.1
parso==0.7.0
pathlib==1.0.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
pre-commit==1.21.0
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.5.0
prompt-toolkit==2.0.10
protobuf==3.14.0
psutil==5.7.2
ptyprocess==0.6.0
py4j==0.10.9.2
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.6.0
pycparser==2.19
pycryptodome==3.9.9
pydocstyle==5.0.2
pyflakes==2.2.0
pyglet==1.4.5
Pygments==2.4.2
pylint==2.5.2
pynvml==8.0.4
pyparsing==2.4.2
pypmml==0.9.11
pyquaternion==0.9.9
pyrsistent==0.15.4
python-dateutil==2.8.0
python-jsonrpc-server==0.3.4
python-language-server==0.33.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
rarfile==3.1
recordio==0.1.7
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rope==0.17.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.3.0
seaborn==0.10.0
Send2Trash==1.5.0
sentencepiece==0.1.85
seqeval==1.2.2
shellcheck-py==0.7.1.1
six==1.15.0
sklearn==0.0
smmap==3.0.5
snowballstemmer==2.0.0
SoundFile==0.10.3.post1
tabulate==0.8.3
tb-nightly==1.15.0a20190801
tb-paddle==0.3.6
tensorboard==2.1.0
tensorboardX==1.8
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.1
typing==3.7.4.3
ujson==1.35
urllib3==1.25.6
virtualenv==16.7.9
visualdl==2.2.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.12.1
xarray==0.16.2
xgboost==1.3.3
xlrd==1.2.0
yapf==0.26.0
zipp==0.6.0